+++ gusto_embedded_v_2025_11_15/src/funcs/contractorPaymentsGetV1ContractorPaymentsContractorPaymentIdPdf.ts
@@ -19,6 +19,10 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import {
+  NotFoundErrorObject,
+  NotFoundErrorObject$inboundSchema,
+} from "../models/errors/notfounderrorobject.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import {
@@ -47,6 +51,7 @@ export function contractorPaymentsGetV1ContractorPaymentsContractorPaymentIdPdf(
 ): APIPromise<
   Result<
     GetV1ContractorPaymentsContractorPaymentIdPdfResponse,
+    | NotFoundErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -72,6 +77,7 @@ async function $do(
   [
     Result<
       GetV1ContractorPaymentsContractorPaymentIdPdfResponse,
+      | NotFoundErrorObject
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -110,7 +116,7 @@ async function $do(
   )(pathParams);
 
   const headers = new Headers(compactMap({
-    Accept: "*/*",
+    Accept: "application/json",
     "X-Gusto-API-Version": encodeSimple(
       "X-Gusto-API-Version",
       payload["X-Gusto-API-Version"],
@@ -172,6 +178,7 @@ async function $do(
 
   const [result] = await M.match<
     GetV1ContractorPaymentsContractorPaymentIdPdfResponse,
+    | NotFoundErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -185,7 +192,8 @@ async function $do(
       200,
       GetV1ContractorPaymentsContractorPaymentIdPdfResponse$inboundSchema,
     ),
-    M.fail([404, "4XX"]),
+    M.jsonErr(404, NotFoundErrorObject$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {