+++ gusto_embedded_v_2025_11_15/src/funcs/externalPayrollsDelete.ts
@@ -19,8 +19,16 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import {
+  NotFoundErrorObject,
+  NotFoundErrorObject$inboundSchema,
+} from "../models/errors/notfounderrorobject.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
+import {
+  UnprocessableEntityError,
+  UnprocessableEntityError$inboundSchema,
+} from "../models/errors/unprocessableentityerror.js";
 import {
   DeleteV1ExternalPayrollRequest,
   DeleteV1ExternalPayrollRequest$outboundSchema,
@@ -47,6 +55,8 @@ export function externalPayrollsDelete(
 ): APIPromise<
   Result<
     DeleteV1ExternalPayrollResponse,
+    | NotFoundErrorObject
+    | UnprocessableEntityError
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -72,6 +82,8 @@ async function $do(
   [
     Result<
       DeleteV1ExternalPayrollResponse,
+      | NotFoundErrorObject
+      | UnprocessableEntityError
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -111,7 +123,7 @@ async function $do(
   )(pathParams);
 
   const headers = new Headers(compactMap({
-    Accept: "*/*",
+    Accept: "application/json",
     "X-Gusto-API-Version": encodeSimple(
       "X-Gusto-API-Version",
       payload["X-Gusto-API-Version"],
@@ -173,6 +185,8 @@ async function $do(
 
   const [result] = await M.match<
     DeleteV1ExternalPayrollResponse,
+    | NotFoundErrorObject
+    | UnprocessableEntityError
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -183,7 +197,9 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(204, DeleteV1ExternalPayrollResponse$inboundSchema),
-    M.fail([404, 422, "4XX"]),
+    M.jsonErr(404, NotFoundErrorObject$inboundSchema),
+    M.jsonErr(422, UnprocessableEntityError$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {