+++ gusto_embedded_v_2025_11_15/src/funcs/externalPayrollsFinalizeTaxLiabilities.ts
@@ -19,8 +19,16 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import {
+  NotFoundErrorObject,
+  NotFoundErrorObject$inboundSchema,
+} from "../models/errors/notfounderrorobject.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
+import {
+  UnprocessableEntityError,
+  UnprocessableEntityError$inboundSchema,
+} from "../models/errors/unprocessableentityerror.js";
 import {
   PutV1TaxLiabilitiesFinishRequest,
   PutV1TaxLiabilitiesFinishRequest$outboundSchema,
@@ -50,6 +58,8 @@ export function externalPayrollsFinalizeTaxLiabilities(
 ): APIPromise<
   Result<
     PutV1TaxLiabilitiesFinishResponse,
+    | NotFoundErrorObject
+    | UnprocessableEntityError
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -75,6 +85,8 @@ async function $do(
   [
     Result<
       PutV1TaxLiabilitiesFinishResponse,
+      | NotFoundErrorObject
+      | UnprocessableEntityError
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -109,7 +121,7 @@ async function $do(
   )(pathParams);
 
   const headers = new Headers(compactMap({
-    Accept: "*/*",
+    Accept: "application/json",
     "X-Gusto-API-Version": encodeSimple(
       "X-Gusto-API-Version",
       payload["X-Gusto-API-Version"],
@@ -171,6 +183,8 @@ async function $do(
 
   const [result] = await M.match<
     PutV1TaxLiabilitiesFinishResponse,
+    | NotFoundErrorObject
+    | UnprocessableEntityError
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -181,7 +195,9 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(202, PutV1TaxLiabilitiesFinishResponse$inboundSchema),
-    M.fail([404, 422, "4XX"]),
+    M.jsonErr(404, NotFoundErrorObject$inboundSchema),
+    M.jsonErr(422, UnprocessableEntityError$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {